Reorganize resource code
