+++ src/vmm/resource/mod.rs
@@ -1,0 +1,31 @@
+use std::path::PathBuf;
+
+use super::ownership::ChangeOwnerError;
+
+pub mod created;
+pub mod moved;
+
+#[derive(Debug)]
+pub enum VmmResourceError {
+    FilesystemError(std::io::Error),
+    MkfifoError(std::io::Error),
+    ChangeOwnerError(ChangeOwnerError),
+    SourcePathMissing(PathBuf),
+}
+
+impl std::error::Error for VmmResourceError {}
+
+impl std::fmt::Display for VmmResourceError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            VmmResourceError::FilesystemError(err) => {
+                write!(f, "A filesystem operation backed by the runtime failed: {err}")
+            }
+            VmmResourceError::MkfifoError(err) => write!(f, "Creating a named pipe via mkfifo failed: {err}"),
+            VmmResourceError::ChangeOwnerError(err) => write!(f, "An ownership change failed: {err}"),
+            VmmResourceError::SourcePathMissing(path) => {
+                write!(f, "The source path of a resource is missing: {}", path.display())
+            }
+        }
+    }
+}
